HU07 Better organize endpoints
