+++ hotel-service/src/main/java/com/playko/hotelservice/service/IHotelService.java
@@ -12,6 +12,4 @@ public interface IHotelService {
     void saveHotel(HotelModel hotelModel);
 
     List<HotelModel> getHotelList(int page, int elementsXpage);
-
-    List<RoomModel> getRooms(Long hotelId, int page, int elementsXpage);
 }
